Filter feeds using pre-built lxml
AWS requires a special build of lxml. Pip-installing it will
silently fail. However, uploading a build that works on Amazon
Linux in the root directory will make it available to
BeautifulSoup. Lxml is thus included in dev-requirements.txt for
local testing but not in requirements.txt, which Now uses to
install dependences on AWS, etc.